Move dim names C++ functions to API
