+++ src/RcppExports.cpp
@@ -235,6 +235,26 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// rray__new_empty_dim_names
+Rcpp::List rray__new_empty_dim_names(int n);
+RcppExport SEXP _rray_rray__new_empty_dim_names(SEXP nSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    rcpp_result_gen = Rcpp::wrap(rray__new_empty_dim_names(n));
+    return rcpp_result_gen;
+END_RCPP
+}
+// rray__dim_names
+Rcpp::List rray__dim_names(const Rcpp::RObject& x);
+RcppExport SEXP _rray_rray__dim_names(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< const Rcpp::RObject& >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(rray__dim_names(x));
+    return rcpp_result_gen;
+END_RCPP
+}
 // rray__dim
 Rcpp::IntegerVector rray__dim(const Rcpp::RObject& x);
 RcppExport SEXP _rray_rray__dim(SEXP xSEXP) {
@@ -362,26 +382,6 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::RObject >::type true_(true_SEXP);
     Rcpp::traits::input_parameter< Rcpp::RObject >::type false_(false_SEXP);
     rcpp_result_gen = Rcpp::wrap(rray__if_else(condition, true_, false_));
-    return rcpp_result_gen;
-END_RCPP
-}
-// rray__new_empty_dim_names
-Rcpp::List rray__new_empty_dim_names(int n);
-RcppExport SEXP _rray_rray__new_empty_dim_names(SEXP nSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::traits::input_parameter< int >::type n(nSEXP);
-    rcpp_result_gen = Rcpp::wrap(rray__new_empty_dim_names(n));
-    return rcpp_result_gen;
-END_RCPP
-}
-// rray__dim_names
-Rcpp::List rray__dim_names(const Rcpp::RObject& x);
-RcppExport SEXP _rray_rray__dim_names(SEXP xSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::traits::input_parameter< const Rcpp::RObject& >::type x(xSEXP);
-    rcpp_result_gen = Rcpp::wrap(rray__dim_names(x));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1058,6 +1058,8 @@ static const R_CallMethodDef CallEntries[] = {
     {"_rray_rray__not_equal", (DL_FUNC) &_rray_rray__not_equal, 2},
     {"_rray_rray__all_equal", (DL_FUNC) &_rray_rray__all_equal, 2},
     {"_rray_rray__any_not_equal", (DL_FUNC) &_rray_rray__any_not_equal, 2},
+    {"_rray_rray__new_empty_dim_names", (DL_FUNC) &_rray_rray__new_empty_dim_names, 1},
+    {"_rray_rray__dim_names", (DL_FUNC) &_rray_rray__dim_names, 1},
     {"_rray_rray__dim", (DL_FUNC) &_rray_rray__dim, 1},
     {"_rray_rray__dim2", (DL_FUNC) &_rray_rray__dim2, 2},
     {"_rray_rray__dims", (DL_FUNC) &_rray_rray__dims, 1},
@@ -1070,8 +1072,6 @@ static const R_CallMethodDef CallEntries[] = {
     {"_rray_rray__any", (DL_FUNC) &_rray_rray__any, 2},
     {"_rray_rray__all", (DL_FUNC) &_rray_rray__all, 2},
     {"_rray_rray__if_else", (DL_FUNC) &_rray_rray__if_else, 3},
-    {"_rray_rray__new_empty_dim_names", (DL_FUNC) &_rray_rray__new_empty_dim_names, 1},
-    {"_rray_rray__dim_names", (DL_FUNC) &_rray_rray__dim_names, 1},
     {"_rray_rray__split", (DL_FUNC) &_rray_rray__split, 2},
     {"_rray_rray__rotate", (DL_FUNC) &_rray_rray__rotate, 4},
     {"_rray_rray__transpose", (DL_FUNC) &_rray_rray__transpose, 2},
